insecure checkbox, move client to a global (ugh), ui tweaks

+++ bundle-addon.sh
@@ -2,6 +2,13 @@
 
 # Bundles all the necessary files into a single `.zip` that Blender understands.
 
+if [[ $# -eq 0 ]] ; then
+    echo 'expected single argument pointing at `core` git directory'
+    echo 'usage: ' $0 ' <core-git-directory>'
+    echo '  where <core-git-directory> is the root directory of https://github.com/blenderfarm/core'
+    exit 0
+fi
+
 rm dist -rf
 mkdir dist
